+++ src/homemaster/benchmarking/browser_demo/__init__.py
@@ -1,0 +1,1 @@
+"""Generic-browser demonstration evaluation support."""